+++ src/hm/hmInverter.h
@@ -255,9 +255,8 @@ class Inverter {
                         if (FLD_T == rec->assign[pos].fieldId) {
                             // temperature is a signed value!
                             rec->record[pos] = (REC_TYP)((int16_t)val) / (REC_TYP)(div);
-                        } else if ((FLD_YT == rec->assign[pos].fieldId)
-                                    && (config->yieldCor != 0)) {
-                            rec->record[pos] = ((REC_TYP)(val) / (REC_TYP)(div)) - ((REC_TYP)config->yieldCor);
+                        } else if (FLD_YT == rec->assign[pos].fieldId) {
+                            rec->record[pos] = ((REC_TYP)(val) / (REC_TYP)(div)) + ((REC_TYP)config->yieldCor[rec->assign[pos].ch]);
                         } else {
                             if ((REC_TYP)(div) > 1)
                                 rec->record[pos] = (REC_TYP)(val) / (REC_TYP)(div);